mirrored view of edit.vue in create.vue, added mixin for shared methods
